# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: Global Football Capital — multi-player football economy platform
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /transfers/market:
    get:
      operationId: getTransferMarket
      tags: [transfers]
      summary: Browse listed players on transfer market
      parameters:
        - name: position
          in: query
          schema:
            type: string
        - name: maxPrice
          in: query
          schema:
            type: integer
        - name: leagueId
          in: query
          schema:
            type: integer
      responses:
        "200":
          description: Market listings
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferListing"
  /transfers/bid:
    post:
      operationId: placeBid
      tags: [transfers]
      summary: Coach bids for a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BidInput"
      responses:
        "200":
          description: Bid placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferBid"
        "400":
          description: Error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /transfers/list-player:
    post:
      operationId: listPlayerForTransfer
      tags: [transfers]
      summary: Coach lists a player for transfer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ListPlayerInput"
      responses:
        "200":
          description: Listed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransferListing"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    TransferListing:
      type: object
      properties:
        id:
          type: integer
        playerId:
          type: integer
        playerName:
          type: string
        position:
          type: string
        overall:
          type: integer
        age:
          type: integer
        fromClubId:
          type: integer
        fromClubName:
          type: string
        askingPrice:
          type: integer
        listedAt:
          type: string
      required: [id, playerId, playerName, position, overall, age, fromClubId, fromClubName, askingPrice, listedAt]
    BidInput:
      type: object
      properties:
        listingId:
          type: integer
        bidAmount:
          type: integer
      required: [listingId, bidAmount]
    TransferBid:
      type: object
      properties:
        id:
          type: integer
        listingId:
          type: integer
        playerName:
          type: string
        bidAmount:
          type: integer
        status:
          type: string
          enum: [pending, accepted, rejected]
        taxAmount:
          type: integer
      required: [id, listingId, playerName, bidAmount, status, taxAmount]
    ListPlayerInput:
      type: object
      properties:
        playerId:
          type: integer
        askingPrice:
          type: integer
      required: [playerId, askingPrice]
...