# lib/api-zod/src/generated/types/transferListing.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Global Football Capital — multi-player football economy platform
 * OpenAPI spec version: 0.2.0
 */

export interface TransferListing {
  id: number;
  playerId: number;
  playerName: string;
  position: string;
  overall: number;
  age: number;
  fromClubId: number;
  fromClubName: string;
  askingPrice: number;
  listedAt: string;
}